tiling and overview for all raster

+++ carteApt/main.cpp
@@ -84,21 +84,25 @@ int main(int argc, char *argv[])
 
                 std::shared_ptr<layerBase> l=kv.second;
 
-                if (l->mResolution!=0.0 && dico.lay4Visu(l->Code()) && !l->Expert()){
+                if (l->mResolution!=0.0 && dico.lay4Visu(l->Code()) && !l->Expert() && l->rasterExist()){
                     std::cout << "work on layer " << l->Code() << std::endl;
                     //std::string aCommand ="gdal raster overview add --external --levels=2,4,8,16 --config COMPRESS=LZW "+ l->getPathTif();
-                    // 1) create COG
+                    // 1) create COG if not already tiled
+                    GDALDataset *pIn= (GDALDataset*) GDALOpen(l->getPathTif().c_str(), GA_ReadOnly);
+                    int b1, b2;
+                    pIn->GetRasterBand(1)->GetBlockSize(&b1,&b2);
+                    GDALClose(pIn);
+                    std::cout << " blocksize " << b1 << " " << b2 << std::endl;
+                    std::string aCommand("");
+
+                    if (b2==1){
                     std::string cogfile =l->getPathTif()+"_cog.tif";
-                    std::string bu ="/media/Data10/Forestimator/bu-"+l->NomFileWithExt();
-                    std::string aCommand="gdal_translate "+ l->getPathTif()+" "+ cogfile +" -co TILED=YES -co COPY_SRC_OVERVIEWS=YES -co COMPRESS=DEFLATE";
+                    std::string bu ="/media/Data10/Forestimator/BU/"+l->NomFileWithExt();
+                    aCommand="gdal_translate "+ l->getPathTif()+" "+ cogfile +" -co TILED=YES -co COPY_SRC_OVERVIEWS=YES -co COMPRESS=DEFLATE";
                     std::cout << aCommand << "\n";
                     if (!globTest){
                     system(aCommand.c_str());
                     }
-                    //if (l->mResolution<9){
-                    //aCommand ="gdal raster overview add --external --levels=2,4,8,16 --config COMPRESS=LZW "+ l->getPathTif();
-
-                    //}
                     // 2) move old file to bu
                     aCommand= "mv "+ l->getPathTif() +" " +bu;
                     std::cout << aCommand << "\n";
@@ -106,11 +110,17 @@ int main(int argc, char *argv[])
                     system(aCommand.c_str());
                     }
                     aCommand= "mv "+ cogfile +" " +l->getPathTif();
+                    std::cout << aCommand << "\n";
                     if (!globTest){
                     system(aCommand.c_str());
                     }
+                    }
                     // 3 compute internal overview
-
+                    aCommand= "gdaladdo -r average -minsize 16 " + l->getPathTif();
+                    std::cout << aCommand << "\n";
+                    if (!globTest){
+                    system(aCommand.c_str());
+                    }
                 }
 
             }
